requirements.txt
streamlit
requests
google-genai
pandas
plotly
pytz